Make admin commands show up on main screen

+++ Computer Science Year-1 Term-2/BigNumberGame/Game.App/GameController.cpp
@@ -72,46 +72,70 @@ void GameController::mainMenuLoggedScreenPrint()
     size_t maxSize = 88;
     bool continueGame = currentUser->level > 1;
     size_t textArraySize = continueGame ? 6 : 5;
+
+    bool isAdmin = currentUser->role == UserRoles::Admin;
+    //Change the textArraySize and textArray based on the user being or not being an admin
+    if (isAdmin) textArraySize += 5;
+
     char** textArray = new char* [textArraySize];
+
+    size_t textArrayIndex = 0;
 
     for (size_t i = 0; i < textArraySize; i++)
     {
         textArray[i] = new char[maxSize];
     }
 
-    //Make text with current level
-    char* currentLevelText = new char[100]
+    //Type out the fact that you are in admin profile
+    if (isAdmin)
     {
-        "Current Level: "
-    };
+        char* loggedAsAdminText = new char[100]
+        {
+            "!!!LOGGED AS ADMIN!!!\n"
+        };
+        /*const char* levelText = ConsoleSystem::parseToString(currentUser->level);
+        strcat(currentLevelText, levelText);
+        strcat(currentLevelText, "\n");*/
+        strcpy(textArray[textArrayIndex++], loggedAsAdminText);
+    }
+
+    //Make text with current level
+    char currentLevelText[100];
+    strcpy(currentLevelText, GlobalConstants::MAINMENU_LEVEL_TEXT);
     const char* levelText = ConsoleSystem::parseToString(currentUser->level);
     strcat(currentLevelText, levelText);
     strcat(currentLevelText, "\n");
-    strcpy(textArray[0], currentLevelText);
+    strcpy(textArray[textArrayIndex++], currentLevelText);
 
     //Make text with current lives
-    char currentLivesText[100] = "Lives: ";
+    char currentLivesText[100];
+    strcpy(currentLivesText, GlobalConstants::MAINMENU_LIVES_TEXT);
     const char* livesText = ConsoleSystem::parseToString(currentUser->lives);
     strcat(currentLivesText, livesText);
     strcat(currentLivesText, "\n");
-    strcpy(textArray[1], currentLivesText);
+    strcpy(textArray[textArrayIndex++], currentLivesText);
 
     //Type out the appropriate things depending on the state of the game of the user
     if (continueGame)
     {
-        strcpy(textArray[2], GlobalConstants::MAINMENU_LOGGED_CONTINUEGAME);
-        strcpy(textArray[3], GlobalConstants::MAINMENU_LOGGED_RESTARTGAME);  
+        strcpy(textArray[textArrayIndex++], GlobalConstants::MAINMENU_LOGGED_CONTINUEGAME);
+        strcpy(textArray[textArrayIndex++], GlobalConstants::MAINMENU_LOGGED_RESTARTGAME);
     }
     else
     {
-        strcpy(textArray[2], GlobalConstants::MAINMENU_LOGGED_STARTGAME);
+        strcpy(textArray[textArrayIndex++], GlobalConstants::MAINMENU_LOGGED_STARTGAME);
     }
-    strcpy(textArray[textArraySize - 2], GlobalConstants::MAINMENU_LOGGED_DELETEOWNACCOUNT);
+    if(!isAdmin) strcpy(textArray[textArrayIndex++], GlobalConstants::MAINMENU_LOGGED_DELETEOWNACCOUNT);
+    else
+    {
+        strcpy(textArray[textArrayIndex++], GlobalConstants::ADMIN_GETINFO);
+        strcpy(textArray[textArrayIndex++], GlobalConstants::ADMIN_EXCLUDEFROMLEADERBOARD);
+        strcpy(textArray[textArrayIndex++], GlobalConstants::ADMIN_DELETEACCOUNT);
+        strcpy(textArray[textArrayIndex++], GlobalConstants::ADMIN_RECOVER);
+        strcpy(textArray[textArrayIndex++], GlobalConstants::ADMIN_ADDHELPER);
+    }
 
-
-    //TODO: Type out the admin privileges
-
-    strcpy(textArray[textArraySize - 1], GlobalConstants::MAINMENU_LOGGED_RETURN_TEXT);
+    strcpy(textArray[textArrayIndex++], GlobalConstants::MAINMENU_LOGGED_RETURN_TEXT);
     //Screen Print
     GameUI::printScreenWithText((const char**)textArray, textArraySize, maxSize);
 
@@ -124,25 +148,30 @@ void GameController::mainMenuLoggedScreenPrint()
 bool GameController::mainMenuLogged()
 {
     mainMenuLoggedScreenPrint();
-
+    bool isAdmin = currentUser->role == UserRoles::Admin;
     while (true)
     { 
         char* selection = new char[10000];
-        std::cin >> selection;
+        char firstChar = (char)std::cin.peek();
+        if (firstChar == '\0' || firstChar == '\n') std::cin.ignore();
+        std::cin.getline(selection, 1000);
+        std::cin.clear();
+
         bool returnToScreen = false;
         bool continueGame = currentUser->level > 1;
-        //If the command is invalid while a game hasnt already been started
         if (strcmp(selection, GlobalConstants::COMMAND_RETURN) == 0)
         {
+            //Unlog the current user
+            currentUser = nullptr;
             //Return to last screen
             return true;
         }
         //User wants to delete their own account, so take them to the confirmation screen!
-        else if (strcmp(selection, GlobalConstants::COMMAND_ACCOUNT_DELETE) == 0)
+        else if (!isAdmin && strcmp(selection, GlobalConstants::COMMAND_ACCOUNT_DELETE) == 0)
         {
             //If the user deletes their own account, then we send them all the way back to the title screen
             returnToScreen = deleteOwnAccountConfirmation();
-
+            currentUser = nullptr;
             if (returnToTitleScreen) return true;
         }
         //Users starts a new game
@@ -157,7 +186,32 @@ bool GameController::mainMenuLogged()
         {
             std::cout << "do something";
         }
-        //If the command is invalid
+        //Get info about a user
+        else if (isAdmin && ConsoleSystem::stringStartsWith(selection, GlobalConstants::COMMAND_ADMIN_GETINFO))
+        {
+            std::cout << "get info about somebody";
+        }
+        //Exclude user from leaderboard
+        else if (isAdmin && ConsoleSystem::stringStartsWith(selection, GlobalConstants::COMMAND_ADMIN_EXCLUDE))
+        {
+            std::cout << "exclude somebody from leaderboards";
+        }
+        //Delete user
+        else if (isAdmin && ConsoleSystem::stringStartsWith(selection, GlobalConstants::COMMAND_ADMIN_DELETE))
+        {
+            std::cout << "delete somebody";
+        }
+        //Recover user
+        else if (isAdmin && ConsoleSystem::stringStartsWith(selection, GlobalConstants::COMMAND_ADMIN_RECOVER))
+        {
+            std::cout << "recover somebody's account";
+        }
+        //Add powerup to admin
+        else if (isAdmin && ConsoleSystem::stringStartsWith(selection, GlobalConstants::COMMAND_ADMIN_ADD))
+        {
+            std::cout << "add a helper";
+        }
+        //If the command is invalid while a game hasnt already been started
         else 
         {
             GameUI::printLineNoBorders(GlobalConstants::COMMAND_INVALID);
@@ -247,7 +301,6 @@ bool GameController::loginUser()
             }
 
             currentUser = FileSystem::getUser(username);
-
             //TODO: If this user is deleted, then print a message to them from the second table of the database
             if (currentUser->isDeleted)
             {
@@ -423,7 +476,7 @@ void GameController::startUpScreenPrint()
 void GameController::startUp()
 {
     //Seed database
-    Seeder::seedDatabase(false);
+    //Seeder::seedDatabase(false);
 
     //Screen print
     startUpScreenPrint();
